chore: move model test files to right package
